requirements.txt
Django==4.2.6
psycopg2-binary==2.9
pdfminer.six
python-docx

# Add other dependencies as needed
